Pushed test name handling forward.
In both `optimize tests` and `record tests`, tests are identified by
their test paths.

Introduced an abstraction in the client to build test paths. For
`optimize tests`, Optimize still access a string that looks like a file
name and builds a default test path. But this is currently not aligned
with what `record tests` do, so something needs to change.

Bazel implementation is now complete, and it demonstrates why this
change was needed in the first place --- test paths reported by `record
tests` should be a sub-path of those used by `optimize tests`