Tweaked the error message
On the read side, when this happens is more complicated, as I think the error only happens when a file exists but not readable. So some kind of uid mixup maybe?

I can't think of a situation where `LAUNCHABLE_SESSION_DIR` helps when read fails. Clearly, write has succeeded first.