A common operator error is that zero tests are passed.
We do not fail here because it's possible that a catastrophic build
failure occur before any tests run, leaving no test reports behind.

OTOH it's also possible during the setup that this simply points to a
wrong location and we end up not collecting any test reports, so it's
worth warning the user.