Add an option for ctest to write regex into split files
CTest has a size limit in the regex engine
(https://gitlab.kitware.com/cmake/cmake/-/blob/2974ea7c3e9cbf9952d6c361bbcecb5027f69605/Source/kwsys/RegularExpression.cxx#L362).
Because of this, we cannot specify a large regex for test subsets.

This change teaches the ctest profile to split the regex and output
files in the specified directory. The caller can opt-in to this new
behavior with --output-regex-files. With this, the CLI writes out the
regex to multiple files, limiting each file size so that it fits to
the CTest regex upperbound.