drop --executable docker option
It may be that no one use the docker option. And the docker image is old, not following up with recent new feature of the cli such as audit and dry-run.

So, how about delete it?

I leave the option as a hidden flag for now, just to see what happens.