Add a root exception handler
By using standalone_mode
(https://click.palletsprojects.com/en/8.0.x/exceptions/), we can get
exceptions from Click without it processing them. This allows us to
catch them and exit with an arbitrary exit code.

This changes the default behavior to exit the CLI always with the exit
code 0, which is a breaking change.

---

While this appears to be a safe change on the first sight, this can
silently disrupt some integrations. Imagine the scenario that, an
integration script is written in a way that if launchable command fails,
it falls back to a full run. With this change, that fallback stops
working, and it can end up reading an empty test subset, and not
executing any test in the worst case.