remove redundant parenthesis
Investigation of a test failure led me to a belief that this code path was redundant.

This part of the code is designed to read `target/maven-status/maven-compiler-plugin/testCompile/default-testCompile/createdFiles.lst`, which is a list of files. The previous 'else' clause, which attempts to put a file name as a class name then create a test path, cannot be right.

This change primarily restores the expected behaviour, that a file that's not a subject of a test shouldn't be in the list of test paths.