Support using submodules as the root repository
Prior to this change, `launchable record build` cannot handle submodules
repository in `--src` option. When a submoduled repository is specified
instead of the submodule parent repository there, the commit collection
code won't be able to open that repository. This is caused by the fact
that the object database of the submodule repository does not reside in
the .git directory, but it's in the parent repository's .git directory.

This change makes it open the submodule repository by letting JGit find
the proper object database location.