Updated based on a customer support request.
Typical maven projects are multi-module, and enumerating modules is not
practical. At least one customer attempted to use '**' but unable to
realize that their shell is expanding it, not our CLI, and got stuck.

Switching to a more realistic example that's ready for copy/paste, and
calling out an attention to the use of quotes to preserve GLOB.