Add a git-log parser
This implements --import-git-log-output option. In some cases, it's hard
to access the git repository from the machine that connects to the
Internet. This makes a workaround that a user can create a file based on
a Git log command (shown in the doc comment), transfer that to the
machine with the Internet connection, and run a commit ingestion.

Tested with one of the demo workspaces.