Teach the raw profile to take a JUnit XML file
Currently the raw profile only takes a JSON file. This change make it
take a JUnit XML file. When the JUnit XML file is used, the CLI needs to
take an option to map a testcase node to a TestPath. This can be
templated based on 'testclass' and 'name'. The mapping option is
required with no default by design to force the users to think about
the alignment with the subset input.