Recorded the result of code spelunking
A customer used `subset --build ...` and hit the error "No saved build
name found". I do belive at some point we revisited this area of "how to
manage complex test session layout" and came to the conclusion that it's
best to let them explicitly deal with sessions, or else they end up
easily creating multiple test sessions when they shouldn't, which is
hard to spot and fix.

What I couldn't figure out is, originally what behaviour was expeected
when `subset --build ...` was run, and whether it indeed required a
locally saved build name. I thought it was possible that changes since
then ended up introducing a bug.

I went back to 917123f0374c80d9cc98750157c64eb1c316e8f3 and verified
that already back then (in 2021), `subset --build` did require a locally
saved build name, and that it matches the given build name. So this is
not a new regression.

This comment records that fact.