Added a test to ensure our bits are Java8 compatible
Renovate wants to bump up dependencies all the time, so it's important
now that we can flag dependencies that are too new.

I originally had animal-sniffer as another test only artifact, but Bazel
still complained that plexus depends on junit yet plexus is not marked
test only. My guess is that the test only bit doesn't flow through
transitive dependencies.

I suppose I could have explicitly marked every intermediate
dependencies, but that didn't seem to be worth the effort, so I
chickened out to an easier path, which is just to remove the test only
flag from junit. It makes this file shorter, which is a plus, too.

The default visibility in the src package was needed in order to make
deploy_jar visible to the test package. I would have expected the
existing public visibility in the java_library would have been
sufficient, but evidently it wasn't. This is a small project, so the
downside of the loose visibility is limited.