Collect head commit, even if it's old
One way we ensure that the CI process integration is working correctly
is to make sure all the commits a build refers to are in our possession.
This is non-trivial to get right especially when `record build` is run
without the commit collection and separate `record commit` is in use.

The intent of filtering by time is to avoid large initial ingestion, so
this change is still consistent with that goal.