Removing logback
The error that prompted this is that slf4j is complaining it's getting
two bindings.

It appears that logback has always been meant as a backend to slf4j API
as the frontend. It's just that in the old version of logback we were
using, that was targetting older slf4j API, so slf4j 2 that we use today
didn't pick up logback as the binding. When we upgraded to a newer
version of logback, slf4j started picking up the binding, hence the
error.

This made me wonder why we bother with loback to begin with. The
reference to logback in CommitIngester suggests it was supposed to be
used by JGit. The current JGit appears to be using slf4j API, so it
doesn't care what backend it is -- perhaps it used to be that it wrote
directly to logback, or maybe chose to bundle logback as the binding?
That'd be a no-no, but hey.

Opting to remove logback, and just go pure slf4j and one of its built-in
bindings. This seems to work OK.