Error check improvement
An experienced customer experiemented with the CLI locally to try out a
new command. Upon seeing the --session option, they assumed they just
need to pass a test session ID, so they ended up invoking something like
`launchable foo bar --session 12345`

That is a good assumption on their part, but it's actually incorrect, as
the ID format we take is more complicated. However, there's no error
check, so the CLI resulted in 404, which misled them.

By adding parse_session, we can make sure the ID given from the command
line is in the correct format, and fail gracefully if not.