Cleaned up a convoluted handling of key=value option parsing
I found out while working on a different issue that key=value get first
parsed into a string representation of Python tuple,
such as "('k', 'v')" then it gets reprocseed into [(k,v)]

It made no sense why we do it this way. The whole point of click is to
process string to typed object conversions within click, and keep the
application code string free.

The said 'different issue' I was working on also involve using a special
"value-less" form of key/value pair option (--branch name as opposed to
--branch repo=name). This use is consistent with how the '--source'
option is defined in `record build`, but it is incompatible with the
KeyValueType's current behaviour of allowing `--option key value`
format.

I pondered how best to proceed, and decided to unsupport this 3 token
version of the KeyValueType handling. This was never used in the
documentation (really, supporting multiple delimiters was never a good
idea IMO), and I figured this is unlikely to be in use in production.

Furthermore, when multiple=True is used, Click wasn't using List, but it
was using Tuple. This meant the type signature of a whole bunch of such
parameters are wrong. I fixed that as well. This meant the default value
can be now `()` (which is immutable) instead of `[]` (which is mutable).
So there's a small gain in the robustness.