Resolving dependency conflicts
Pipfile specifies a set of constraints on dependencies, and therefore as
newer versions of dependencies release, how they resolve will change.

My understanding is that, as a result of this, some time ago our
`pipenv lock` stopped functioning. This has been bothering me for a while,
but finally I bit the bullet and resolved it.

To prevent this problem further, I'm also committing Pipfile.lock, which
stores the results of dependency resolutions. This way, we can guarantee
that we always have a set of versions that satisfies constraints.

As noted in the comment, it is possible that this can create different
problems around platform dependent stuff. But so far, we don't seem to
have that issue.