Redone the dockerfile
The immediate motivation for the change was from a prospect who had the
opsec policy to avoid root in containers. That in turn revealed the
problem that the current container is built with venv, a rather odd
choise for a container!

It turns out that we probably didn't choose venv because we were clever.
Rather, the base image was ubuntu and it disabled/discouraged global pip
package installation as it collides with APT.

I could have ignored that warning anyway, but at this point it seems to
me that Ubuntu is not helping, it's hurting. So I switched to
python:3.x-slim, which appears to be built with the `pip install
--global xyz` mindset.