Handle error when java -version returns error
While /usr/bin/java might exist on users macOS system, when no JDK or JRE is installed, it typically act as a placeholder or symlink meant to guide users toward installing Java. So in such case, when we call `java -version` it return an error instead of the version, which we need to handle.